# api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Restaurant operations platform API",
    "description": "Orders at the table, tickets live in the kitchen, a bill at the end.",
    "license": {
      "name": ""
    },
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/floor": {
      "get": {
        "tags": [
          "orders"
        ],
        "summary": "The floor, in the order it is walked.",
        "description": "# Errors\n\nReturns `401` if nobody is signed in, `403` if the caller is not a waiter,\nand `503` if the database is unavailable.",
        "operationId": "floor",
        "responses": {
          "200": {
            "description": "Every live table and its occupancy. Waiters only.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FloorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Nobody is signed in.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "403": {
            "description": "Not a waiter.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/kitchen/tickets": {
      "get": {
        "tags": [
          "orders"
        ],
        "summary": "The kitchen queue, oldest first.",
        "description": "# Errors\n\nReturns `401` if nobody is signed in, `403` if the caller is not a chef, and\n`503` if the database is unavailable.",
        "operationId": "kitchen_tickets",
        "responses": {
          "200": {
            "description": "Every queued or ready ticket, oldest first. Chefs only.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/KitchenResponse"
                }
              }
            }
          },
          "401": {
            "description": "Nobody is signed in.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "403": {
            "description": "Not a chef.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/rounds/{id}/served": {
      "post": {
        "tags": [
          "orders"
        ],
        "summary": "Marks a whole ticket as having reached the table.",
        "description": "The one refusal here that is not a repository operation's own. It reads the\nticket first and refuses unless it is `ready` at that moment, because\n\"somebody has already carried this out\" is what the waiter needs to be told,\nand a line level message about one dish would not say it.\n\nInside the loop it marks every dish that is `ready` and skips one that is\nalready `served` or was cancelled rather than treating either as a conflict.\nTwo waiters serving the same ticket at the same instant is not a mistake\nworth stopping the second one over; only the precheck refuses, and only\nbecause by then the ticket is no longer ready.\n\n# Errors\n\nReturns `409 round_not_ready` if the ticket is not waiting to be carried\nout, `404` if there is no such ticket, `401` if nobody is signed in, and\n`403` if the caller is not a waiter.",
        "operationId": "mark_round_served",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "The ticket that reached the table.",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The ticket after the write. Waiters only.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrderRoundDto"
                }
              }
            }
          },
          "401": {
            "description": "Nobody is signed in.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "403": {
            "description": "Not a waiter.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "404": {
            "description": "No such ticket.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "409": {
            "description": "That ticket is not waiting to be carried out.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        }
      }
    },
    "/api/visits": {
      "post": {
        "tags": [
          "orders"
        ],
        "summary": "Seats a party and opens their bill, in one transaction.",
        "description": "Both or neither. A visit with no bill would be a table that could take an\norder nobody could be charged for, and it would be discovered at the end of\nthe meal.\n\n# Errors\n\nReturns `409 table_occupied` if somebody else opened that table first, `400`\nif the table is archived or not this restaurant's, `401` if nobody is signed\nin, and `403` if the caller is not a waiter.",
        "operationId": "open_visit",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OpenVisitRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "The visit and its bill. Waiters only.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OpenVisitResponse"
                }
              }
            }
          },
          "400": {
            "description": "No such live table.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "401": {
            "description": "Nobody is signed in.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "403": {
            "description": "Not a waiter.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "409": {
            "description": "That table already has a party at it.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ErrorBody": {
        "type": "object",
        "description": "The body every failed request returns.",
        "required": [
          "error",
          "message"
        ],
        "properties": {
          "error": {
            "type": "string",
            "description": "A stable machine readable code, safe to branch on in the client.\n\nA refused conflict names which conflict it was, such as\n`table_occupied` or `bill_has_unserved_lines`, rather than the blanket\nword `conflict`. The closed list is\n[`ConflictKind`](crate::domain::error::ConflictKind).",
            "example": "not_found"
          },
          "fields": {
            "type": [
              "object",
              "null"
            ],
            "description": "What is wrong with each named field, when the problem is a form rather\nthan the request as a whole.\n\nAbsent from every response that has no field level problem, which is\nmost of them, so nothing that was already reading this body changes.\nEach value is one of a closed set of codes: `already_taken`,\n`too_short`, `too_long`, `invalid_format`, `unknown_country`,\n`not_in_catalogue`, `incorrect`, `required`.",
            "additionalProperties": {
              "type": "string"
            },
            "propertyNames": {
              "type": "string"
            },
            "example": {
              "email": "already_taken"
            }
          },
          "message": {
            "type": "string",
            "description": "A human readable sentence. Never contains internal detail.\n\nEnglish, and for a log. The web app renders a translation of\n[`Self::error`] instead, because the API is a data API and knows nothing\nabout what language anybody reads.",
            "example": "The requested resource does not exist."
          }
        }
      },
      "FloorResponse": {
        "type": "object",
        "description": "Every table in the restaurant, grouped by section.",
        "required": [
          "sections"
        ],
        "properties": {
          "sections": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/FloorSectionDto"
            },
            "description": "The sections, in displayed order, each with its tables."
          }
        }
      },
...
      "KitchenResponse": {
        "type": "object",
        "description": "Everything the kitchen still has work on.",
        "required": [
          "tickets",
          "serverTime"
        ],
        "properties": {
          "serverTime": {
            "type": "string",
            "format": "date-time",
            "description": "What the server's clock reads, at the moment this answer was built.\n\nThe screen subtracts this from its own clock and corrects every age it\ndraws by the difference. A kitchen tablet whose clock is twenty minutes\nfast would otherwise show every ticket as twenty minutes late, and the\none number a chef acts on would be the one number nobody could trust."
          },
          "tickets": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/KitchenTicketDto"
            },
            "description": "The tickets, oldest first."
          }
        }
      },
...
      "OrderRoundDto": {
        "type": "object",
        "description": "One ticket with its dishes.",
        "required": [
          "id",
          "visitId",
          "sequenceNo",
          "status",
          "sentAt",
          "lines"
        ],
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "description": "Which ticket this is."
          },
          "lines": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/OrderLineDto"
            },
            "description": "Every dish on it, oldest first."
          },
          "readyAt": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time",
            "description": "When the last dish came off the pass."
          },
          "sentAt": {
            "type": "string",
            "format": "date-time",
            "description": "When it reached the kitchen."
          },
          "sequenceNo": {
            "type": "integer",
            "format": "int32",
            "description": "Its number within the visit, starting at one."
          },
          "servedAt": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time",
            "description": "When the last dish reached the table."
          },
          "status": {
            "$ref": "#/components/schemas/RoundStatusDto",
            "description": "Where the whole ticket has got to, recomputed from its dishes."
          },
          "visitId": {
            "type": "string",
            "format": "uuid",
            "description": "Which visit it belongs to."
          }
        }
      },
...
    }
  },
...
}